backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
pydantic>=2.9.0
pydantic-settings>=2.5.0
requests>=2.32.0
google-generativeai>=0.8.0
python-multipart>=0.0.12

# Gmail API for OAuth email sending
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0

# Job scheduling
apscheduler>=3.10.4

# Database packages
sqlalchemy>=2.0.36
alembic>=1.13.0
psycopg2-binary>=2.9.9

